- update Grade and kotlin versions - update code
