+++ build.gradle.kts
@@ -1,8 +1,8 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id("com.android.application").version("8.1.1").apply(false)
-    id("com.android.library").version("8.1.1").apply(false)
-    kotlin("android").version("1.9.0").apply(false)
+    id("com.android.application").version("8.1.2").apply(false)
+    id("com.android.library").version("8.1.2").apply(false)
+    kotlin("android").version("1.9.10").apply(false)
     id("io.gitlab.arturbosch.detekt").version("1.23.1").apply(false)
 }
