+++ filepickerlibrary/src/main/java/com/nareshchocha/filepickerlibrary/ui/activitys/VideoCaptureActivity.kt
@@ -199,7 +199,6 @@ internal class VideoCaptureActivity : AppCompatActivity() {
                     PackageManager.PackageInfoFlags.of(PackageManager.GET_PERMISSIONS.toLong()),
                 )
             } else {
-                @Suppress("DEPRECATION")
                 context.packageManager.getPackageInfo(
                     context.packageName,
                     PackageManager.GET_PERMISSIONS,
